Trazendo novas funcionalidades do infra.

See merge request almatec/alma-infra!54

+++ Alma.Core/IIdNome.cs
@@ -4,7 +4,7 @@
     {
     }
 
-    public interface IId : IId<int>
+    public interface IId : IId<long>
     {
     }
